Refactor: make Cards list semi-reusable
