+++ src/components/explore/CardList.tsx
@@ -1,6 +1,12 @@
 import React, { FC } from 'react'
 import { createStyles, makeStyles, Theme } from '@material-ui/core/styles'
 
+import { SwatchOnly } from 'components/legend'
+import { useParams, useRouteMatch } from 'react-router-dom'
+import { RouteMatch, TonsWithAddl } from './types'
+import { getUniqueInstances } from './utils'
+import { CustomCard } from './CustomCard'
+
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
     // TODO: masonry someday:
@@ -18,11 +24,54 @@ const useStyles = makeStyles((theme: Theme) =>
   })
 )
 
-// TODO: https://react-window.now.sh/#/examples/list/fixed-size
-// Simple grid wrapper designed for Card children
-export const CardList: FC = (props) => {
+export const CardListWrap: FC = (props) => {
   const { children } = props
   const classes = useStyles()
 
   return <div className={classes.root}>{children}</div>
+}
+
+// TODO: https://react-window.now.sh/#/examples/list/fixed-size
+// Simple grid wrapper designed for Card children
+export const CardList: FC<{ data: TonsWithAddl[]; tableName: string }> = (
+  props
+) => {
+  const { data, tableName } = props
+  const classes = useStyles()
+  const { field, value } = useParams<RouteMatch & { value: string }>()
+  const { url } = useRouteMatch()
+
+  return (
+    <div className={classes.root}>
+      {data.map((row) => {
+        const uniqueInstances = getUniqueInstances(field, row, value)
+        const nameOrLang = row.name || row.Language
+
+        return (
+          <CustomCard
+            key={nameOrLang}
+            intro={value}
+            title={tableName === 'Language' ? row.Endonym : nameOrLang}
+            uniqueInstances={uniqueInstances}
+            url={`${url}/${nameOrLang}`}
+            // TODO: use and refactor SwatchOrFlagOrIcon for icon prop
+            icon={
+              <>
+                {row['icon-color'] && (
+                  <SwatchOnly backgroundColor={row['icon-color']} />
+                )}
+                {row.src_image && (
+                  <img
+                    style={{ height: '0.8em', marginRight: '0.25em' }}
+                    src={row.src_image[0].url}
+                    alt={nameOrLang}
+                  />
+                )}
+              </>
+            }
+          />
+        )
+      })}
+    </div>
+  )
 }
